requirement.txt
asgiref==3.3.4
Django==3.1.11
djangorestframework==3.11.2
pytz==2021.1
sqlparse==0.4.1
